[#1384] neofs-cli: move `storagegroup` command to a separate package
